Move impoersation methods to their own helper for better segregation
