+++ app/helpers/application_helper.rb
@@ -9,22 +9,6 @@ module ApplicationHelper
     end
   end
 
-  def impersonating_agent(user)
-    session[:impersonating_agent_id] = user.id
-  end
-
-  def delete_impersonating_agent
-    session.delete :impersonating_agent_id
-  end
-
-  def impersonating_agent_id
-    session[:impersonating_agent_id]
-  end
-
-  def student_impersonation?
-    impersonating_agent_id.present?
-  end
-
   # Adding current user role to page class
   def body_class
     classes = []
